fixed up the navbar issue and added create player edit player now i need to fix physical metrics

+++ backend/scouting_service/src/Scoutings/Core/Application/Features/PhysicalMetrics/GetAllPhysicalMetrics/GetAllPhysicalMetricsResponse.cs
@@ -7,7 +7,7 @@ public class GetAllPhysicalMetricsResponse
     public int? FatPercentage { get; set; }
     public int? BenchPressWeight { get; set; }
     public int? SquatWeight { get; set; }
-    public int? SprintSpeed { get; set; }
+    public decimal? SprintSpeed { get; set; }
     public int? Weight { get; set; }
     public int? Height { get; set; }
     public int? Wingspan { get; set; }
