+++ backend/scouting_service/src/Scoutings/Core/Application/Features/Players/GetAllPlayers/GetAllPlayersResponse.cs
@@ -22,7 +22,7 @@ public class PhysicalMetricResponse
     public int? FatPercentage { get; set; }
     public int? BenchPressWeight { get; set; }
     public int? SquatWeight { get; set; }
-    public int? SprintSpeed { get; set; }
+    public decimal? SprintSpeed { get; set; }
     public int? Weight { get; set; }
     public int? Height { get; set; }
     public int? Wingspan { get; set; }
